updated migration file

+++ db/migrate/20220313040703_add_column_post_recipes.rb
@@ -1,6 +1,8 @@
 class AddColumnPostRecipes < ActiveRecord::Migration[5.2]
   def change
-    add_column :post_recipes, :image_id, :string
-    remove_column :post_recipes, :"image", :string
+    if table_exists? :post_recipes
+      add_column :post_recipes, :image_id, :string
+      remove_column :post_recipes, :"image", :string
+    end
   end
 end
