+++ db/migrate/20220322171238_post_recipes.rb
@@ -1,5 +1,5 @@
 class PostRecipes < ActiveRecord::Migration[5.2]
   def change
-    drop_table :post_recipes
+    drop_table :post_recipes if table_exists? :post_recipes
   end
 end
