+++ db/migrate/20220322171320_steps.rb
@@ -1,5 +1,5 @@
 class Steps < ActiveRecord::Migration[5.2]
   def change
-    drop_table :steps
+    drop_table :steps if table_exists? :steps
   end
 end
